refactor: move hero section to sections
